+++ src/pending_tasks/api/views.py
@@ -7,13 +7,15 @@ from django.shortcuts import get_object_or_404
 from django.urls import reverse
 from django.utils.translation import gettext_lazy as _
 
-from pending_tasks.api.model_viewset_without_create import (
-    AuthAllPermBaseObjectWithoutCreate,
-)
-from rest_framework import mixins
+from django_filters.rest_framework import DjangoFilterBackend
+from djgentelella.permission_management import AllPermissionByAction
+from rest_framework import mixins, status
+from rest_framework.authentication import TokenAuthentication, SessionAuthentication
 from rest_framework.decorators import action
+from rest_framework.filters import SearchFilter, OrderingFilter
+from rest_framework.pagination import LimitOffsetPagination
 from rest_framework.response import Response
-from rest_framework import status
+from rest_framework.viewsets import GenericViewSet
 
 from auth_and_perms.models import Profile
 from laboratory.models import OrganizationStructure
@@ -25,16 +27,25 @@ from pending_tasks.api.serializers import (
     PendingTaskSerializer,
     PendingTaskListSerializer,
     PendingTaskValidateSerializer,
-    ProfileValidateSerializer,
-    CurrentStatusValidateSerializer,
-    NewStatusValidateSerializer,
     PendingTaskGetValuesSerializer,
 )
 
 logger = logging.getLogger("organilab")
 
 
-class PendingTaskViewSet(mixins.CreateModelMixin, AuthAllPermBaseObjectWithoutCreate):
+class PendingTaskViewSet(
+    mixins.CreateModelMixin,
+    mixins.RetrieveModelMixin,
+    mixins.UpdateModelMixin,
+    mixins.DestroyModelMixin,
+    mixins.ListModelMixin,
+    GenericViewSet,
+):
+    authentication_classes = (TokenAuthentication, SessionAuthentication)
+    permission_classes = (AllPermissionByAction,)
+    pagination_class = LimitOffsetPagination
+    filter_backends = (DjangoFilterBackend, SearchFilter, OrderingFilter)
+
     serializer_class = {
         "list": PendingTaskListSerializer,
         "create": PendingTaskValidateSerializer,
@@ -53,9 +64,6 @@ class PendingTaskViewSet(mixins.CreateModelMixin, AuthAllPermBaseObjectWithoutCr
         "update": ["pending_tasks.change_pendingtask"],
         "partial_update": ["pending_tasks.change_pendingtask"],
         "destroy": ["pending_tasks.delete_pendingtask"],
-        "task_assign": ["pending_tasks.change_pendingtask"],
-        "task_unassign": ["pending_tasks.change_pendingtask"],
-        "updated_task_status": ["pending_tasks.change_pendingtask"],
         "archive_finished": ["pending_tasks.change_pendingtask"],
     }
 
@@ -64,6 +72,27 @@ class PendingTaskViewSet(mixins.CreateModelMixin, AuthAllPermBaseObjectWithoutCr
     filterset_class = filterset.PendingTaskFilterSet
     ordering_fields = ["creation_date"]
     ordering = ("-creation_date",)
+
+    def get_serializer_class(self):
+        if self.action in self.serializer_class:
+            return self.serializer_class[self.action]
+        return super().get_serializer_class()
+
+    def list(self, request, *args, **kwargs):
+        queryset = self.filter_queryset(self.get_queryset())
+        data = self.paginate_queryset(queryset)
+        response = {
+            "data": data,
+            "recordsTotal": self.queryset.count(),
+            "recordsFiltered": queryset.count(),
+            "draw": request.GET.get("draw", 1),
+        }
+        return Response(self.get_serializer(response).data)
+
+    @action(detail=True, methods=["get"])
+    def get_values_for_update(self, request, *args, **kwargs):
+        instance = self.get_object()
+        return Response(self.get_serializer(instance).data)
 
     def perform_create(self, serializer):
         org = get_object_or_404(OrganizationStructure, pk=self.kwargs.get("org_pk"))
@@ -89,7 +118,6 @@ class PendingTaskViewSet(mixins.CreateModelMixin, AuthAllPermBaseObjectWithoutCr
             profilepermission__rol__in=instance.rols.all()
         ).values_list("user_id", flat=True)
         users_to_notify.update(role_user_ids)
-
         users_to_notify.discard(self.request.user.pk)
 
         User = get_user_model()
@@ -100,59 +128,12 @@ class PendingTaskViewSet(mixins.CreateModelMixin, AuthAllPermBaseObjectWithoutCr
         queryset = super().filter_queryset(queryset)
         profile = self.request.user.profile
         rols = profile.profilepermission_set.all().values_list("rol", flat=True)
-        queryset = (
+        return (
             queryset.filter(
                 Q(profile=profile) | Q(rols__in=rols) | Q(created_by=self.request.user)
             )
             .filter(is_archived=False)
             .distinct()
-        )
-        return queryset
-
-    def _get_task_and_data(self, request):
-        task = self.get_object()
-        data = {"profile": request.user.profile.id}
-        data.update(request.data)
-        return task, data
-
-    def _execute_task_action(self, request, serializer_class, apply_changes):
-        task, data = self._get_task_and_data(request)
-        serializer = serializer_class(data=data, context={"task": task})
-        if serializer.is_valid():
-            changed_data = apply_changes(task, serializer)
-            task.save()
-            organilab_logentry(request.user, task, CHANGE, changed_data=changed_data)
-            return Response(PendingTaskSerializer(task).data, status=status.HTTP_200_OK)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-
-    @action(detail=True)
-    def task_assign(self, request, *args, **kwargs):
-        def apply_changes(task, serializer):
-            task.profile = serializer.validated_data["profile"]
-            return ["profile"]
-
-        return self._execute_task_action(
-            request, ProfileValidateSerializer, apply_changes
-        )
-
-    @action(detail=True)
-    def task_unassign(self, request, *args, **kwargs):
-        def apply_changes(task, serializer):
-            task.profile = None
-            return ["profile"]
-
-        return self._execute_task_action(
-            request, CurrentStatusValidateSerializer, apply_changes
-        )
-
-    @action(detail=True, methods=["patch"])
-    def updated_task_status(self, request, *args, **kwargs):
-        def apply_changes(task, serializer):
-            task.status = serializer.validated_data.get("status")
-            return ["status"]
-
-        return self._execute_task_action(
-            request, NewStatusValidateSerializer, apply_changes
         )
 
     @action(detail=False, methods=["post"], url_path="archive_finished")
@@ -160,10 +141,7 @@ class PendingTaskViewSet(mixins.CreateModelMixin, AuthAllPermBaseObjectWithoutCr
         profile = request.user.profile
         rols = profile.profilepermission_set.all().values_list("rol", flat=True)
         updated = (
-            PendingTask.objects.filter(
-                status=PendingTask.FINISHED,
-                is_archived=False,
-            )
+            PendingTask.objects.filter(status=PendingTask.FINISHED, is_archived=False)
             .filter(Q(profile=profile) | Q(rols__in=rols) | Q(created_by=request.user))
             .distinct()
             .update(is_archived=True)
