+++ src/pending_tasks/templates/tasks/tasks-view.html
@@ -64,7 +64,7 @@
 	{% include 'gentelella/blocks/modal_template.html' with form=create_form id="create_obj_modal" title=create_obj_tittle form_id="create_obj_form" url=list_obj_url principal_class="modal-lg" %}
 
 	{% url "pending_tasks:api-pending_tasks-detail" org_pk 0 as detail_obj_url %}
-	{% trans 'Update task' as update_obj_tittle %}
+	{% trans 'Edit task' as update_obj_tittle %}
 	{% include 'gentelella/blocks/modal_template.html' with form=update_form id="update_obj_modal" title=update_obj_tittle form_id="update_obj_form" url=detail_obj_url principal_class="modal-lg" %}
 	{% include 'gentelella/blocks/modal_template_delete.html' with form=delete_form id="delete_obj_modal" form_id="delete_obj_form" url=detail_obj_url %}
