+++ src/pending_tasks/tests.py
@@ -254,94 +254,6 @@ class PendingTaskAPIDetailTest(PendingTaskSetUpMixin, TestCase):
         self.assertEqual(self.task.description, "Updated description")
 
 
-class PendingTaskAssignTest(PendingTaskSetUpMixin, TestCase):
-
-    def setUp(self):
-        super().setUp()
-        self.task = PendingTask.objects.create(
-            organization=self.org,
-            created_by=self.user,
-            description="Unassigned task",
-            status=PendingTask.PENDING,
-        )
-        self.task.rols.add(self.role)
-
-    def test_task_assign(self):
-        url = reverse(
-            "pending_tasks:pending_tasks-task-assign",
-            kwargs={"pk": self.task.pk},
-        )
-        response = self.client.get(url)
-        self.assertEqual(response.status_code, 200)
-        self.task.refresh_from_db()
-        self.assertEqual(self.task.profile, self.user.profile)
-
-    def test_task_assign_already_assigned(self):
-        self.task.profile = self.user.profile
-        self.task.save()
-        url = reverse(
-            "pending_tasks:pending_tasks-task-assign",
-            kwargs={"pk": self.task.pk},
-        )
-        response = self.client.get(url)
-        self.assertEqual(response.status_code, 400)
-
-    def test_task_assign_unauthenticated(self):
-        self.client.logout()
-        url = reverse(
-            "pending_tasks:pending_tasks-task-assign",
-            kwargs={"pk": self.task.pk},
-        )
-        response = self.client.get(url)
-        self.assertIn(response.status_code, [401, 403])
-
-
-class PendingTaskUnassignTest(PendingTaskSetUpMixin, TestCase):
-
-    def setUp(self):
-        super().setUp()
-        self.task = PendingTask.objects.create(
-            organization=self.org,
-            created_by=self.user,
-            description="Assigned task",
-            status=PendingTask.PENDING,
-            profile=self.user.profile,
-        )
-        self.task.rols.add(self.role)
-
-    def test_task_unassign(self):
-        url = reverse(
-            "pending_tasks:pending_tasks-task-unassign",
-            kwargs={"pk": self.task.pk},
-        )
-        response = self.client.get(url)
-        self.assertEqual(response.status_code, 200)
-        self.task.refresh_from_db()
-        self.assertIsNone(self.task.profile)
-
-    def test_task_unassign_not_assigned_to_user(self):
-        other_user = create_user_with_profile("otherunassign")
-        self.task.profile = other_user.profile
-        self.task.save()
-        url = reverse(
-            "pending_tasks:pending_tasks-task-unassign",
-            kwargs={"pk": self.task.pk},
-        )
-        response = self.client.get(url)
-        # 404 because filter_queryset excludes tasks not visible to the user
-        self.assertEqual(response.status_code, 404)
-
-    def test_task_unassign_when_not_pending(self):
-        self.task.status = PendingTask.IN_PROCESS
-        self.task.save()
-        url = reverse(
-            "pending_tasks:pending_tasks-task-unassign",
-            kwargs={"pk": self.task.pk},
-        )
-        response = self.client.get(url)
-        self.assertEqual(response.status_code, 400)
-
-
 class PendingTaskUpdateStatusTest(PendingTaskSetUpMixin, TestCase):
 
     def setUp(self):
